requirements.txt
boto3==1.10.12
botocore==1.13.12
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==5.1.2
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.6
certifi==2019.9.11
chardet==3.0.4
idna==2.8
jsonpickle==1.2
python-digitalocean==1.14.0
requests==2.22.0
PyCryptodome==3.9.4
inputimeout==1.0.4
progressbar2==3.47.0